# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Complete Booking System api information",
    "contact": {
      "name": "",
      "url": "",
      "email": ""
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Server 1"
    }
  ],
  "paths": {
    "/booking/reserve": {
      "post": {
        "tags": [
          "Booking"
        ],
        "description": "Reserve a product before payment.",
        "responses": {
          "200": {
            "description": "Product reserved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string"
                  }
                },
                "required": [
                  "productId"
                ]
              }
            }
          }
        }
      }
    },
    "/booking/pay": {
      "post": {
        "tags": [
          "Booking"
        ],
        "description": "Confirm booking by processing a mock payment.",
        "responses": {
          "200": {
            "description": "Booking confirmed after successful mock payment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/responseObject"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": {
                    "type": "string"
                  },
                  "sessionId": {
                    "type": "string"
                  }
                },
                "required": [
                  "bookingId",
                  "sessionId"
                ]
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "responseObject": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {}
          }
        }
      },
      "responseArray": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "array",
            "example": [],
            "items": {}
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "basicAuth": []
    }
  ]
}